separated integration tests into auth-stack(pre-auth & auth) and main-tab(post-auth)
