+++ SmartClothingApp/__test__/__integration-test__/__main-tab-test__/__main-navigation-test/HomeToProfile.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import HomeScreen from '../../../src/screens/HomeScreen/index.jsx';
-import ProfileScreen from '../../../src/screens/Profile/index.jsx';
-import configureStore from '../../../src/store.js';
+import HomeScreen from '../../../../src/screens/HomeScreen/index.jsx';
+import ProfileScreen from '../../../../src/screens/Profile/index.jsx';
+import configureStore from '../../../../src/store.js';
 import {Provider as StoreProvider } from 'react-redux'; 
 import {PaperProvider}  from "react-native-paper";
 import { render } from '@testing-library/react-native';
@@ -16,7 +16,7 @@ import { act } from 'react-test-renderer';
 const Stack = createStackNavigator();
 
 // Mock Firebase Authentication
-jest.mock('../../../firebaseConfig.js', () => ({
+jest.mock('../../../../firebaseConfig.js', () => ({
     auth: {
       loginWithEmail: jest.fn(() => Promise.resolve()),
       startLoginWithEmail: jest.fn(() => Promise.resolve()),
@@ -43,7 +43,7 @@ jest.mock('firebase/auth', () => ({
     getDatabase: jest.fn(),
   }))
 
-jest.mock('../../../src/utils/localStorage.js', () => ({
+jest.mock('../../../../src/utils/localStorage.js', () => ({
     AsyncStorage: jest.fn(),
     storeUID: jest.fn(),
     getUID: jest.fn(),
@@ -67,10 +67,10 @@ jest.mock('firebase/firestore', () => ({
     }),
   }))
 
-jest.mock('react-native-vector-icons/MaterialIcons', () => require('../../__mocks__/react-native-vector-icons.js').MaterialIcons);
-  jest.mock('react-native-vector-icons/FontAwesome5', () => require('../../__mocks__/react-native-vector-icons.js').FontAwesome5);
-  jest.mock('@shopify/react-native-skia', () => require('../../__mocks__/@shopify__react-native-skia.js'));
-  jest.mock('../../../src/components/visualizations/ActivityRings/Ring.jsx', () => {
+jest.mock('react-native-vector-icons/MaterialIcons', () => require('../../../__mocks__/react-native-vector-icons.js').MaterialIcons);
+  jest.mock('react-native-vector-icons/FontAwesome5', () => require('../../../__mocks__/react-native-vector-icons.js').FontAwesome5);
+  jest.mock('@shopify/react-native-skia', () => require('../../../__mocks__/@shopify__react-native-skia.js'));
+  jest.mock('../../../../src/components/visualizations/ActivityRings/Ring.jsx', () => {
     return jest.fn(({ ring, center, strokeWidth, scale }) => (
       <div>
         Mock Ring Component - {ring.size}, {center.x}, {center.y}, {strokeWidth}, {scale}
@@ -116,7 +116,20 @@ jest.mock('react-native-vector-icons/MaterialIcons', () => require('../../__mock
 
 
 describe('Home to Profile Navigation', () => {
+
   let store;
+
+  beforeEach(() => {
+    // Prevents 'wrap act()' console log warning 
+    jest.spyOn(console, 'error').mockImplementation((message) => {
+      if (message.includes('Warning: An update to')) {
+        return;
+      }
+      console.error(message);
+    });
+
+    jest.useFakeTimers();
+  });
 
   it('Navigates from Home to Profile screen', async () => {
     store = configureStore();
