+++ SmartClothingApp/__test__/__integration-test__/__main-tab-test__/__main-screen-test__/SettingsScreen.test.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import { render, fireEvent } from '@testing-library/react-native';
-import SettingsScreen from '../src/screens/Settings/index.jsx';
+import SettingsScreen from '../../../../src/screens/Settings/index.jsx';
 import { Provider as ReduxProvider } from 'react-redux';
 import { Provider as PaperProvider } from 'react-native-paper';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import configureStore from '../src/store.js';
+import configureStore from '../../../../src/store.js';
 
 // Mock Firebase Authentication
-jest.mock('../firebaseConfig.js', () => ({
+jest.mock('../../../../firebaseConfig.js', () => ({
     auth: {
       loginWithEmail: jest.fn(() => Promise.resolve()),
       startLoginWithEmail: jest.fn(() => Promise.resolve()),
@@ -29,7 +29,7 @@ jest.mock('../firebaseConfig.js', () => ({
     },
   }));
 
-  jest.mock('../src/utils/localStorage.js', () => ({
+  jest.mock('../../../../src/utils/localStorage.js', () => ({
     AsyncStorage: jest.fn(),
     storeUID: jest.fn(),
     getUID: jest.fn(),
@@ -73,50 +73,61 @@ jest.mock('firebase/firestore', () => ({
 const Stack = createStackNavigator();
 
 describe('SettingsScreen', () => {
-    let store;
-    let component;
-    let instance;
+  let store;
+  let component;
+  let instance;
 
-    beforeEach(() => {
-        store = configureStore();
-        component = renderer.create(
-          <ReduxProvider store={store}>
-            <PaperProvider>
-              <NavigationContainer>
-                <Stack.Navigator>
-                  <Stack.Screen
-                    name="Settings"
-                    component={SettingsScreen}
-                    initialParams={{ previousScreenTitle: 'Test Title' }}
-                  />
-                </Stack.Navigator>
-              </NavigationContainer>
-            </PaperProvider>
-          </ReduxProvider>
-        );
-        instance = component.root;
-      });
+  beforeEach(() => {
 
-    it('renders correctly', () => {
-        expect(component.toJSON()).toMatchSnapshot(); 
+    // Prevents console log warning 
+    jest.spyOn(console, 'error').mockImplementation((message) => {
+      if (message.includes('Warning: An update to')) {
+        return;
+      }
+      console.error(message);
     });
-    
-    it('opens update email modal on button press', () => {
-        const updateEmailButton = instance.findByProps({ title: "UPDATE EMAIL" });
-        fireEvent.press(updateEmailButton);
-        // Expect the state for the update email modal to be true
-    });
-    
-    it('opens change password modal on button press', () => {
-        const changePasswordButton = instance.findByProps({ title: "CHANGE PASSWORD" });
-        fireEvent.press(changePasswordButton);
-        // Expect the state for the change password modal to be true
-    });
-    
-    it('opens delete account modal on button press', () => {
-        const deleteAccountButton = instance.findByProps({ title: "DELETE ACCOUNT" });
-        fireEvent.press(deleteAccountButton);
-        // Expect the state for the delete account modal to be true
-    });
-  // Add more tests for navigation, modal interactions, and Redux actions
+
+    jest.useFakeTimers();
+
+    store = configureStore();
+    component = renderer.create(
+      <ReduxProvider store={store}>
+        <PaperProvider>
+          <NavigationContainer>
+            <Stack.Navigator>
+              <Stack.Screen
+                name="Settings"
+                component={SettingsScreen}
+                initialParams={{ previousScreenTitle: 'Test Title' }}
+              />
+            </Stack.Navigator>
+          </NavigationContainer>
+        </PaperProvider>
+      </ReduxProvider>
+    );
+    instance = component.root;
+  });
+
+  it('renders correctly', () => {
+      expect(component.toJSON()).toMatchSnapshot(); 
+  });
+  
+  it('opens update email modal on button press', () => {
+      const updateEmailButton = instance.findByProps({ title: "UPDATE EMAIL" });
+      fireEvent.press(updateEmailButton);
+      // Expect the state for the update email modal to be true
+  });
+  
+  it('opens change password modal on button press', () => {
+      const changePasswordButton = instance.findByProps({ title: "CHANGE PASSWORD" });
+      fireEvent.press(changePasswordButton);
+      // Expect the state for the change password modal to be true
+  });
+  
+  it('opens delete account modal on button press', () => {
+      const deleteAccountButton = instance.findByProps({ title: "DELETE ACCOUNT" });
+      fireEvent.press(deleteAccountButton);
+      // Expect the state for the delete account modal to be true
+  });
+// Add more tests for navigation, modal interactions, and Redux actions
 });
